Verify Board
We may want to add more conditions

+++ leon-go/src/main/scala/go/util/Range.scala
@@ -3,8 +3,8 @@ import leon.collection.List
 
 object Range {
 
-  def to(from: Int, to: Int): List[Int] = if (from > to) List() else from :: this.to(from + 1, to)
+  def to(from: BigInt, to: BigInt): List[BigInt] = if (from > to) List() else from :: this.to(from + 1, to)
 
-  def until(from: Int, to: Int): List[Int] = if (from >= to) List() else from :: until(from + 1, to)
+  def until(from: BigInt, to: BigInt): List[BigInt] = if (from >= to) List() else from :: until(from + 1, to)
 
 }
